requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# LangChain & AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-groq>=0.0.1
openai>=1.7.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Optional but recommended
httpx>=0.26.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic>=0.4.27
unstructured>=0.11.0
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# LangChain & AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-groq>=0.0.1
openai>=1.7.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Optional but recommended
httpx>=0.26.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic>=0.4.27
unstructured>=0.11.0
